Add some methods for Matrix Factorization
There is a py file just for matrix factorization
methods. There should be two in there; one is an
off the shelf method called surprise SVD that is
recommended for the class. The implementation is
very, very basic and has been tested very little.

There is also a method pretty much canned from the
last set designed to minimize the mean regularized
square error. This will likely be tweaked with before
the final submission.