Implement Bias Method
With some helper functions, there is the bias method. 4 of the helpers
are gradient functions, 2 of them are gradient functions without
bias terms, the next is the method to get error with the bias terms,
the one after that is the method to get the error without the
bias terms, train model function trains te model with bias terms, and
the Vtrain model function trains the model without the bias terms.

The main function runs these following steps: 1. Run SVD for Y with
bias terms, where the return values include the transposed U and V,
2. Run SVD for V without bias terms, which returns the transposed A,
3. Find the projection through taking the product of the transposed
first two columns of A and U/V, 4. The projection is compressed where
each term is divided by the maximum of its row, 5. and we get the
error and then we return the projected U and projected V.